The state machine has been layed out, and the maze detection code is in place.  Code just needs to be written to do something with it.
Added in ray casting tests to detect when an enemy can be seen and the script will try to shoot it.  Currently, it needs to be updated to aim first, then shoot.

I also added 2 sets of detectors to try to detect the maze around the tank.  The code to decide which path to take needs to be written.