Did a quick hack in TankDamage.js to prevent the tank flying off the map, however this glitches the controls, so we'll have to look at this again.
Added a GameOverScreen prefab that is spawned when the game is over, rather than switching scenes.  This will allow us to very easily get the scores, instead of trying to figure out how to pass data from 1 scene to the next.

Modified the TankAI script with more comments and logic.

Deleted the GameOver scene (not sure why it is still here, but thats fine).

Modified the GUI to display a FPS counter in the bottom left corner, and now the player tank score backgrounds will flash in time with the tank lights, just because it seemed extra cool to do.