Updated the game controller to use a min function on the fps counter as it was giving errors when the frames per second number would overflow an integer.
Updated the script details for shooting and AI for all 3 tanks to be the same.

Updated the AI pathfinding abilities to prevent them from spinning in circles.

Modified the TankDamage respawn code to make the tank spawn in the middle of the spawn point by adding an offset to their position to compensate for the tank graphic not being dead center.